Move constans file
