+++ src/Transaction.js
@@ -1,5 +1,5 @@
 const Tx = require('./Transaction/Tx.js');
-const Constants = require('./Constants.js');
+const Constants = require('./Config/Constants.js');
 // Below import for intellisense and type support on jsdoc
 const Wallet = require('./Wallet.js'); //eslint-disable-line
